+++ uc-820.js
@@ -211,4 +211,4 @@
     .version-info p {
       margin: 4px 0;
     }
-  `,R([(0,i.MZ)({attribute:!1})],M.prototype,"hass",void 0),M=R([(0,i.EM)("ultra-about-tab")],M)},4541:(e,t,o)=>{o.d(t,{x:()=>a});const a="3.1.0-beta19"}}]);
+  `,R([(0,i.MZ)({attribute:!1})],M.prototype,"hass",void 0),M=R([(0,i.EM)("ultra-about-tab")],M)},4541:(e,t,o)=>{o.d(t,{x:()=>a});const a="3.1.0-beta20"}}]);